LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/inline-block-expected.txt

abc

def

FAIL Anchor selection descent into inline blocks. assert_equals: expected 200 but got 150
